refactor(persistence): move SqliteWorkflowPersistence into runkon-flow (Phase 4 step 4.3)
The canonical implementation now lives in runkon-flow/src/persistence_sqlite.rs,
gated on the rusqlite feature alongside the existing in-memory variant. The
file is fully self-contained: row mappers, column constants, and the small
gate / json helpers all live next to the struct, so a harness only needs to
flip the rusqlite feature on to get a production-ready persistence backend.

conductor-core's workflow/persistence_sqlite.rs becomes a thin re-export plus
the integration-test suite that exercises the implementation against
conductor's actual schema (migrations + AgentManager + WorkflowManager
fixtures). Existing call sites that import SqliteWorkflowPersistence from
crate::workflow continue to work unchanged.

Open() is now a public convenience constructor (previously cfg(test)) since
sensible PRAGMA defaults — WAL, foreign keys, busy_timeout — apply equally
to any harness using SQLite. Production callers continue to share a
connection via from_shared_connection.

cargo test --workspace, cargo clippy --workspace --all-targets -- -D warnings,
and cargo fmt --all --check are all green.
