+++ runkon-flow/src/lib.rs
@@ -11,6 +11,8 @@ pub mod helpers;
 pub mod output_schema;
 #[cfg(any(test, feature = "test-utils"))]
 pub mod persistence_memory;
+#[cfg(feature = "rusqlite")]
+pub mod persistence_sqlite;
 pub mod prompt_builder;
 pub mod status;
 #[cfg(any(test, feature = "test-utils"))]
